setup - Batch commit, Refactoring base view structure, Configuring basic tailwind config,
