+++ package.json
@@ -14,7 +14,7 @@
         "axios": "^1.7.4",
         "laravel-vite-plugin": "^1.0",
         "postcss": "^8.4.32",
-        "tailwindcss": "^3.4.0",
+        "tailwindcss": "^3.4.13",
         "vite": "^5.0",
         "vue": "^3.3.13"
     }
