+++ resources/js/Layouts/Footer.vue
@@ -1,0 +1,11 @@
+<script setup>
+defineProps({
+    on: Boolean,
+});
+</script>
+
+<template>
+    <footer class="py-16 text-center text-xl text-black dark:text-white/70 bg-red-600">
+        <h2>Footer</h2>
+    </footer>
+</template>
